+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -----------------------------------------------------------------------------------------------
 ## v1.0.0 unreleased
 
+### Added
+  - `Params` can be constructed with `numberOfRegions`, `numberOfBoundaryRegions` and `numberOfCarriers`
+  - `Params` can be constructed directly from problem specific parameter structs
+
 ### Changed
   - parameter files are replaced by parameter structs with explicit parameter access: `p = parameter_set(); p.foo` to access parameter `foo`;
     you can specify in the examples which `parameter_set` is used.